BUG: Duplicating Procurement Resource forms - I believe this was caused by a combination of factors: * setting 'initial' on formset creation populated the display fine, but if nothing was edited on the page before saving, those initial forms were ignored by the formset save * if something was set on one or more of the forms, then only those forms would be saved. * if an error occurred to stop a complete 6 form save, then initial would create another 7 forms along with whatever was saved. Solution: PR objects are created on first load of the P-Form now (much like Organs are too). This means no initial data, only the loading of the instance data.
BUG: Admin interface would not delete the PR forms from an Organ Admin interface. This was due to a fault in the replacement save_formset logic in compare/admin.py. Using formset.save() over a loop with each instance being saved in the formset has fixed this.

P-Form save processing has now been split into the 6 separate forms/ets so that each section of the form will validate and save independant of the other. This works because all objects are created ahead of the first form load, so there is no waiting for reference keys on initial save

T-Form Organ Allocation has been amended to use the same principles of initial creation over "initial" population of the formset factory.

There is still a bit of a logic trap for when Organ Allocations are changed, but not so much of one that it causes the program to crash.